Move fto tree package
