+++ server/src/main/java/server/BUILD
@@ -10,7 +10,7 @@ java_library(
     ],
     deps = [
         "//server/src/main/java/server/analysis",
-        "//server/src/main/java/server/bazel",
+        "//server/src/main/java/server/bazel/tree",
         "//server/src/main/java/server/buildifier",
         "//server/src/main/java/server/completion",
         "//server/src/main/java/server/highlighting",
